Refactor frontend
