--- src/ReconNess.Web/ClientApp/src/components/agent/AgentCreatePage.vue
@@ -1,27 +1,0 @@
-﻿<template>
-  <div>
-    <h3>New Agent</h3>
-    <agent-form v-bind:parentAgent="newAgent"></agent-form>
-  </div>
-</template>
-
-<script>
-  import AgentForm from './ui/AgentForm'
-
-  export default {
-    name: 'AgentCreatePage',
-    components: {
-      AgentForm
-    },    
-    data: () => {
-      return {      
-        newAgent: {},
-      }
-    }  
-  }
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-
-</style>
