+++ src/ReconNess.Web/ClientApp/src/components/home/ReferenceAndResource.vue
