+++ src/ReconNess.Web/ClientApp/src/components/subdomain/SubdomainTagAgents.vue
@@ -18,7 +18,7 @@
             <td class="w-25">{{ agent.categories.join(', ') }}</td>
             <td class="w-25">{{ agent.lastRun | formatDate('YYYY-MM-DD') }}</td>
             <td class="w-25">
-              <button class="btn btn-primary ml-2" v-on:click="onRunAgent(agent)" v-if="!agent.isRunning" :disabled="agentRunning != null && agentRunning.name != agent.name">Run</button>
+              <button class="btn btn-primary ml-2" v-on:click="onRunAgent(agent)" v-if="!agent.isRunning" :disabled="disabledCanRun(agent)">Run</button>
               <button class="btn btn-danger ml-2" v-on:click="onStopAgent(agent)" v-if="agent.isRunning">Stop</button>
               <button class="btn btn-dark ml-2" v-on:click="showTerminalModal = !showTerminalModal" v-if="agent.isRunning">Terminal</button>
               <button class="btn btn-dark ml-2" v-on:click="showLogModal = !showLogModal" v-if="agent.isRunning">Logs</button>
@@ -75,16 +75,19 @@
   import { Terminal } from 'xterm'
 
   export default {
-    name: 'TargetAgents', 
+    name: 'SubdomainTagAgents', 
     props: {
-      parentAgents: {
+      agents: {
         type: Array,
+        required: true
+      },
+      subdomain: {
+        type: Object,
         required: true
       }
     },
     data() {
       return { 
-        agents: [],
         showTerminalModal: false,
         showLogModal: false,
         term: null,
@@ -93,28 +96,28 @@
       }    
     },    
     async mounted() {
-      this.agents = this.parentAgents || []
-            
-      this.agents.map(agent => {
-        const channel = `${this.$route.params.targetName}_${agent.name}`
-        this.$connection.on(channel, (message) => {
-          if (message === "Agent stopped!" || message === "Agent done!")
-          {
-            agent.isRunning = false;
-            this.agentRunning = null
-          }
+      if (this.agents.length > 0) {
+        this.agents.map(agent => {
+          const channel = `${this.$route.params.targetName}_${this.$route.params.subdomain}_${agent.name}`
+          this.$connection.on(channel, (message) => {
+            if (message === "Agent stopped!" || message === "Agent done!") {
+              agent.isRunning = false;
+              this.agentRunning = null
+            }
 
-          if (this.term !== null) {
-            this.term.writeln(message)
-          }
-        });
+            if (this.term !== null) {
+              this.term.writeln(message)
+            }
+          });
 
-        this.$connection.on("logs_" + channel, (message) => {          
-          if (this.termLog !== null) {
-            this.termLog.writeln(message)
-          }
-        });
-      })   
+          this.$connection.on("logs_" + channel, (message) => {
+
+            if (this.termLog !== null) {
+              this.termLog.writeln(message)
+            }
+          });
+        })
+      }
     },    
     methods: {  
       async onRunAgent(agent) {
@@ -127,11 +130,13 @@
         this.agentRunning = agent
 
         const target = this.$route.params.targetName
+        const subdomainName = this.$route.params.subdomain
         const agentName = agent.name      
 
         await this.$api.create('agents/run', {
           agent: agentName,
-          target: target
+          target: target,
+          subdomain: subdomainName
         })      
       },
       async onStopAgent(agent) { 
@@ -142,12 +147,20 @@
         this.agentRunning = null
 
         const target = this.$route.params.targetName
+        const subdomainName = this.$route.params.subdomain
         const agentName = agent.name
 
         await this.$api.create('agents/stop', {
           agent: agentName,
-          target: target
+          target: target,
+          subdomain: subdomainName
         })      
+      },
+      disabledCanRun(agent) {
+        const anotherAgentIsRunning = this.agentRunning != null && this.agentRunning.name !== agent.name
+        const needToBeAlive = agent.onlyIfIsAlive === true && this.subdomain.isAlive !== true
+
+        return needToBeAlive || anotherAgentIsRunning
       },
       startTransitionModal() {      
         this.$refs.backdrop.classList.toggle("d-block");
@@ -159,7 +172,7 @@
           })
           this.term.open(document.getElementById('terminal'))
         }       
-      },
+      },      
       endTransitionModal() {
         this.$refs.backdrop.classList.toggle("show");
         if (this.$refs.modal !== undefined) {
