+++ src/ReconNess.Web/ClientApp/src/components/subdomain/SubdomainTagServices.vue
@@ -25,20 +25,12 @@
 
 <script>
   export default {
-    name: 'SubdomainServices',
+    name: 'SubdomainTagServices',
     props: {
-      parentServices: {
+      services: {
         type: Array,
         required: true
       }
-    },
-    data: () => {
-      return {
-        services: []        
-      }
-    },
-    async mounted() {
-      this.services = this.parentServices || []
     }
   }
 </script>
