+++ src/ReconNess.Web/ClientApp/src/views/agent/AgentDebugPage.vue
